Changed parameter in print_int function and updated prototype in header file

+++ print_int.c
@@ -6,7 +6,7 @@
  * @n: number to be printed
  */
 
-void print_int(int n)
+void print_int(long int n)
 {
 	/* checks if number has a negative sign and prints it */
 	if (n < 0)
